Extract go-to item rows to common and reuse them in related items panel
JournalGoToItemRow and ScrapEntryGoToItemRow move from GoToSimple to
common/itemRows (together with their base GoToItemRow) and now render
the related-items panel as well, instead of a near-duplicate
implementation. The title fallback to a notes excerpt (for log book
entries) lives in the shared row now.
