+++ app/src/components/overview/goto/GoToSimple.tsx
@@ -1,21 +1,13 @@
 import { IEntity } from "../../../serverApi/IEntity";
 import { IJournal } from "../../../serverApi/IJournal";
 import { OverviewList } from "../overviewList/OverviewList";
-import { GoToItemRow } from "./GoToItemRow";
+import { GoToItemRow } from "../../common/itemRows/GoToItemRow";
+import { JournalGoToItemRow } from "../../common/itemRows/JournalGoToItemRow";
+import { ScrapEntryGoToItemRow } from "../../common/itemRows/ScrapEntryGoToItemRow";
 import { IScrapEntry } from "../../../serverApi/IScrapEntry";
 import { knownQueryParams } from "../../common/actions/searchParamHooks";
-import { JournalIcon } from "../journals/JournalIcon";
-import { IconStyle } from "../../common/IconStyle";
-import { Icon } from "../../common/Icon";
-import Check from "@mui/icons-material/Check";
-import Notes from "@mui/icons-material/Notes";
 import SearchOutlined from "@mui/icons-material/SearchOutlined";
 import { useGoToNavigationItems } from "./useGoToNavigationItems";
-import { ActionFactory } from "../../common/actions/ActionFactory";
-import { ActionIconButton } from "../../common/actions/ActionIconButton";
-import { DeviceWidth, useDeviceWidth } from "../../common/useDeviceWidth";
-import { isTypeThatCanShowAddEntryRow } from "../../../util/journalUtils";
-import { JournalType } from "../../../serverApi/JournalType";
 import { useEffect } from "react";
 
 const emptyListItemId = "empty-list-item-id";
@@ -81,82 +73,20 @@ export const GoToSimple: React.FC<{
         <JournalGoToItemRow
           journal={entity as IJournal}
           hasFocus={hasFocus}
-          onClick={onClick ?? (() => {})}
+          onClick={onClick}
         />
       );
     }
 
     return (
       <ScrapEntryGoToItemRow
-        journal={
-          goto.journalsForEntries.find(
-            (j) => j.id === (entity as IScrapEntry).parentId,
-          )!
-        }
+        journal={goto.journalsForEntries.find(
+          (j) => j.id === (entity as IScrapEntry).parentId,
+        )}
         scrapEntry={entity as IScrapEntry}
         hasFocus={hasFocus}
-        onClick={onClick ?? (() => {})}
+        onClick={onClick}
       />
     );
   }
-};
-
-const ScrapEntryGoToItemRow: React.FC<{
-  scrapEntry: IScrapEntry;
-  journal: IJournal;
-  hasFocus: boolean;
-  onClick: () => void;
-}> = ({ scrapEntry, journal, hasFocus, onClick }) => {
-  return (
-    <GoToItemRow
-      icon={
-        <Icon style={IconStyle.Small}>
-          {scrapEntry.scrapType === "List" ? <Check /> : <Notes />}
-        </Icon>
-      }
-      url={`/journals/details/${scrapEntry.parentId}?${knownQueryParams.selectedItemId}=${scrapEntry.id}`}
-      hasFocus={hasFocus}
-      onClick={onClick}
-    >
-      <span style={{ display: "flex", alignItems: "center" }}>
-        {`${scrapEntry.title || scrapEntry.id}`}
-        <span
-          style={{ fontSize: "smaller", color: "initial", paddingLeft: "10px" }}
-        >
-          {journal.name}
-        </span>
-      </span>
-    </GoToItemRow>
-  );
-};
-
-const JournalGoToItemRow: React.FC<{
-  journal: IJournal;
-  hasFocus: boolean;
-  onClick: () => void;
-}> = ({ journal, hasFocus, onClick }) => {
-  const deviceWidth = useDeviceWidth();
-
-  return (
-    <GoToItemRow
-      url={`/journals/details/${journal.id}`}
-      onClick={onClick}
-      hasFocus={hasFocus}
-      icon={<JournalIcon journal={journal} iconStyle={IconStyle.Small} />}
-      renderAtEnd={() => {
-        return journal.type !== JournalType.Scraps ? (
-          <ActionIconButton
-            action={
-              deviceWidth === DeviceWidth.Normal &&
-              isTypeThatCanShowAddEntryRow(journal.type)
-                ? ActionFactory.goToJournal(journal.id ?? "", false)
-                : ActionFactory.addEntry(journal, false, () => {}, true)
-            }
-          />
-        ) : null;
-      }}
-    >
-      {`${journal.name}`}
-    </GoToItemRow>
-  );
 };
